chore: move type building logic out of FTLProcessor
fixes #2360
