+++ jvm-runtime/ftl-runtime/common/deployment/src/main/java/xyz/block/ftl/deployment/FtlProcessor.java
@@ -1,17 +1,12 @@
 package xyz.block.ftl.deployment;
 
 import java.io.IOException;
-import java.lang.reflect.Modifier;
 import java.nio.charset.StandardCharsets;
 import java.nio.file.Files;
 import java.nio.file.Path;
 import java.nio.file.attribute.PosixFilePermission;
-import java.time.Instant;
-import java.time.OffsetDateTime;
-import java.time.ZonedDateTime;
 import java.util.ArrayList;
 import java.util.EnumSet;
-import java.util.HashMap;
 import java.util.HashSet;
 import java.util.List;
 import java.util.Map;
@@ -23,8 +18,6 @@ import java.util.stream.Collectors;
 import org.jboss.jandex.AnnotationInstance;
 import org.jboss.jandex.AnnotationTarget;
 import org.jboss.jandex.ArrayType;
-import org.jboss.jandex.ClassInfo;
-import org.jboss.jandex.ClassType;
 import org.jboss.jandex.DotName;
 import org.jboss.jandex.MethodInfo;
 import org.jboss.jandex.PrimitiveType;
@@ -37,13 +30,10 @@ import org.jboss.resteasy.reactive.server.mapping.URITemplate;
 import org.jboss.resteasy.reactive.server.processor.scanning.MethodScanner;
 import org.jetbrains.annotations.NotNull;
 
-import com.fasterxml.jackson.annotation.JsonAlias;
-import com.fasterxml.jackson.databind.JsonNode;
 import com.fasterxml.jackson.databind.ObjectMapper;
 
 import io.quarkus.agroal.spi.JdbcDataSourceBuildItem;
 import io.quarkus.arc.deployment.AdditionalBeanBuildItem;
-import io.quarkus.arc.processor.DotNames;
 import io.quarkus.deployment.Capabilities;
 import io.quarkus.deployment.Capability;
 import io.quarkus.deployment.annotations.BuildProducer;
@@ -72,7 +62,6 @@ import io.quarkus.vertx.http.runtime.VertxHttpRecorder;
 import xyz.block.ftl.Config;
 import xyz.block.ftl.Cron;
 import xyz.block.ftl.Export;
-import xyz.block.ftl.GeneratedRef;
 import xyz.block.ftl.LeaseClient;
 import xyz.block.ftl.Retry;
 import xyz.block.ftl.Secret;
@@ -91,30 +80,21 @@ import xyz.block.ftl.runtime.builtin.HttpResponse;
 import xyz.block.ftl.runtime.config.FTLConfigSource;
 import xyz.block.ftl.runtime.http.FTLHttpHandler;
 import xyz.block.ftl.v1.CallRequest;
-import xyz.block.ftl.v1.schema.Any;
 import xyz.block.ftl.v1.schema.Array;
-import xyz.block.ftl.v1.schema.Bool;
-import xyz.block.ftl.v1.schema.Bytes;
-import xyz.block.ftl.v1.schema.Data;
 import xyz.block.ftl.v1.schema.Database;
 import xyz.block.ftl.v1.schema.Decl;
-import xyz.block.ftl.v1.schema.Field;
 import xyz.block.ftl.v1.schema.Float;
 import xyz.block.ftl.v1.schema.IngressPathComponent;
 import xyz.block.ftl.v1.schema.IngressPathLiteral;
 import xyz.block.ftl.v1.schema.IngressPathParameter;
-import xyz.block.ftl.v1.schema.Int;
 import xyz.block.ftl.v1.schema.Metadata;
-import xyz.block.ftl.v1.schema.MetadataAlias;
 import xyz.block.ftl.v1.schema.MetadataCalls;
 import xyz.block.ftl.v1.schema.MetadataCronJob;
 import xyz.block.ftl.v1.schema.MetadataIngress;
 import xyz.block.ftl.v1.schema.MetadataRetry;
 import xyz.block.ftl.v1.schema.MetadataSubscriber;
-import xyz.block.ftl.v1.schema.Module;
 import xyz.block.ftl.v1.schema.Optional;
 import xyz.block.ftl.v1.schema.Ref;
-import xyz.block.ftl.v1.schema.Time;
 import xyz.block.ftl.v1.schema.Type;
 import xyz.block.ftl.v1.schema.Unit;
 
@@ -123,21 +103,15 @@ class FtlProcessor {
     private static final Logger log = Logger.getLogger(FtlProcessor.class);
 
     private static final String SCHEMA_OUT = "schema.pb";
+
+    public static final String BUILTIN = "builtin";
     public static final DotName EXPORT = DotName.createSimple(Export.class);
     public static final DotName VERB = DotName.createSimple(Verb.class);
     public static final DotName CRON = DotName.createSimple(Cron.class);
     public static final DotName SUBSCRIPTION = DotName.createSimple(Subscription.class);
-    public static final String BUILTIN = "builtin";
-    public static final DotName CONSUMER = DotName.createSimple(Consumer.class);
     public static final DotName SECRET = DotName.createSimple(Secret.class);
     public static final DotName CONFIG = DotName.createSimple(Config.class);
-    public static final DotName OFFSET_DATE_TIME = DotName.createSimple(OffsetDateTime.class.getName());
-    public static final DotName GENERATED_REF = DotName.createSimple(GeneratedRef.class);
     public static final DotName LEASE_CLIENT = DotName.createSimple(LeaseClient.class);
-    public static final DotName INSTANT = DotName.createSimple(Instant.class);
-    public static final DotName ZONED_DATE_TIME = DotName.createSimple(ZonedDateTime.class);
-    public static final DotName NOT_NULL = DotName.createSimple(NotNull.class);
-    public static final DotName JSON_NODE = DotName.createSimple(JsonNode.class.getName());
 
     @BuildStep
     BindableServiceBuildItem verbService() {
@@ -225,11 +199,10 @@ class FtlProcessor {
             BuildProducer<SystemPropertyBuildItem> systemPropProducer,
             BuildProducer<GeneratedResourceBuildItem> generatedResourceBuildItemBuildProducer) throws Exception {
         String moduleName = moduleNameBuildItem.getModuleName();
-        Module.Builder moduleBuilder = Module.newBuilder()
-                .setName(moduleName)
-                .setBuiltin(false);
-        Map<TypeKey, ExistingRef> dataElements = new HashMap<>();
-        ExtractionContext extractionContext = new ExtractionContext(moduleName, index, recorder, moduleBuilder, dataElements,
+
+        SchemaBuilder schemaBuilder = new SchemaBuilder(index.getComputingIndex(), moduleName);
+
+        ExtractionContext extractionContext = new ExtractionContext(moduleName, index, recorder, schemaBuilder,
                 new HashSet<>(), new HashSet<>(), topics.getTopics(), verbClients.getVerbClients());
         var beans = AdditionalBeanBuildItem.builder().setUnremovable();
 
@@ -254,7 +227,7 @@ class FtlProcessor {
                 systemPropProducer.produce(new SystemPropertyBuildItem(
                         "quarkus.datasource." + ds.getName() + ".credentials-provider-name", FTLDatasourceCredentials.NAME));
             }
-            moduleBuilder.addDecls(
+            schemaBuilder.addDecls(
                     Decl.newBuilder().setDatabase(
                             Database.newBuilder().setType("postgres").setName(sanitisedName))
                             .build());
@@ -264,15 +237,15 @@ class FtlProcessor {
 
         //register all the topics we are defining in the module definition
         for (var topic : topics.getTopics().values()) {
-            extractionContext.moduleBuilder.addDecls(Decl.newBuilder().setTopic(xyz.block.ftl.v1.schema.Topic.newBuilder()
+            extractionContext.schemaBuilder.addDecls(Decl.newBuilder().setTopic(xyz.block.ftl.v1.schema.Topic.newBuilder()
                     .setExport(topic.exported())
                     .setName(topic.topicName())
-                    .setEvent(buildType(extractionContext, topic.eventType(), topic.exported())).build()));
+                    .setEvent(schemaBuilder.buildType(topic.eventType(), topic.exported())).build()).build());
         }
 
         handleVerbAnnotations(index, beans, extractionContext);
         handleCronAnnotations(index, beans, extractionContext);
-        handleSubscriptionAnnotations(index, subscriptionMetaAnnotationsBuildItem, moduleName, moduleBuilder, extractionContext,
+        handleSubscriptionAnnotations(index, subscriptionMetaAnnotationsBuildItem, moduleName, extractionContext,
                 beans);
 
         //TODO: make this composable so it is not just one big method, build items should contribute to the schema
@@ -351,14 +324,14 @@ class FtlProcessor {
                     .setIngress(ingressBuilder
                             .build())
                     .build();
-            Type requestTypeParam = buildType(extractionContext, bodyParamType, true);
-            Type responseTypeParam = buildType(extractionContext, endpoint.getMethodInfo().returnType(), true);
+            Type requestTypeParam = schemaBuilder.buildType(bodyParamType, true);
+            Type responseTypeParam = schemaBuilder.buildType(endpoint.getMethodInfo().returnType(), true);
             Type stringType = Type.newBuilder().setString(xyz.block.ftl.v1.schema.String.newBuilder().build()).build();
             Type pathParamType = Type.newBuilder()
                     .setMap(xyz.block.ftl.v1.schema.Map.newBuilder().setKey(stringType)
                             .setValue(stringType))
                     .build();
-            moduleBuilder
+            schemaBuilder
                     .addDecls(Decl.newBuilder().setVerb(xyz.block.ftl.v1.schema.Verb.newBuilder()
                             .addMetadata(ingressMetadata)
                             .setName(verbName)
@@ -383,7 +356,7 @@ class FtlProcessor {
 
         Path output = outputTargetBuildItem.getOutputDirectory().resolve(SCHEMA_OUT);
         try (var out = Files.newOutputStream(output)) {
-            moduleBuilder.build().writeTo(out);
+            schemaBuilder.writeTo(out);
         }
 
         output = outputTargetBuildItem.getOutputDirectory().resolve("main");
@@ -415,7 +388,7 @@ class FtlProcessor {
 
     private void handleSubscriptionAnnotations(CombinedIndexBuildItem index,
             SubscriptionMetaAnnotationsBuildItem subscriptionMetaAnnotationsBuildItem, String moduleName,
-            Module.Builder moduleBuilder, ExtractionContext extractionContext, AdditionalBeanBuildItem.Builder beans) {
+            ExtractionContext extractionContext, AdditionalBeanBuildItem.Builder beans) {
         for (var subscription : index.getIndex().getAnnotations(SUBSCRIPTION)) {
             var info = SubscriptionMetaAnnotationsBuildItem.fromJandex(subscription, moduleName);
             if (subscription.target().kind() != AnnotationTarget.Kind.METHOD) {
@@ -423,7 +396,7 @@ class FtlProcessor {
             }
             var method = subscription.target().asMethod();
             String className = method.declaringClass().name().toString();
-            generateSubscription(moduleBuilder, extractionContext, beans, method, className, info);
+            generateSubscription(extractionContext, beans, method, className, info);
         }
         for (var metaSub : subscriptionMetaAnnotationsBuildItem.getAnnotations().entrySet()) {
             for (var subscription : index.getIndex().getAnnotations(metaSub.getKey())) {
@@ -432,7 +405,7 @@ class FtlProcessor {
                     continue;
                 }
                 var method = subscription.target().asMethod();
-                generateSubscription(moduleBuilder, extractionContext, beans, method,
+                generateSubscription(extractionContext, beans, method,
                         method.declaringClass().name().toString(),
                         metaSub.getValue());
             }
@@ -453,12 +426,13 @@ class FtlProcessor {
         }
     }
 
-    private void generateSubscription(Module.Builder moduleBuilder, ExtractionContext extractionContext,
+    private void generateSubscription(ExtractionContext extractionContext,
             AdditionalBeanBuildItem.Builder beans, MethodInfo method, String className,
             SubscriptionMetaAnnotationsBuildItem.SubscriptionAnnotation info) {
         beans.addBeanClass(className);
-        moduleBuilder.addDecls(Decl.newBuilder().setSubscription(xyz.block.ftl.v1.schema.Subscription.newBuilder()
-                .setName(info.name()).setTopic(Ref.newBuilder().setName(info.topic()).setModule(info.module()).build()))
+        extractionContext.schemaBuilder.addDecls(Decl.newBuilder()
+                .setSubscription(xyz.block.ftl.v1.schema.Subscription.newBuilder()
+                        .setName(info.name()).setTopic(Ref.newBuilder().setName(info.topic()).setModule(info.module()).build()))
                 .build());
         handleVerbMethod(extractionContext, method, className, false, BodyType.REQUIRED, (builder -> {
             builder.addMetadata(Metadata.newBuilder().setSubscriber(MetadataSubscriber.newBuilder().setName(info.name())));
@@ -494,8 +468,8 @@ class FtlProcessor {
                     String name = param.annotation(Secret.class).value().asString();
                     paramMappers.add(new VerbRegistry.SecretSupplier(name, paramType));
                     if (!context.knownSecrets.contains(name)) {
-                        context.moduleBuilder.addDecls(Decl.newBuilder().setSecret(xyz.block.ftl.v1.schema.Secret.newBuilder()
-                                .setType(buildType(context, param.type(), false)).setName(name)));
+                        context.schemaBuilder.addDecls(Decl.newBuilder().setSecret(xyz.block.ftl.v1.schema.Secret.newBuilder()
+                                .setType(context.schemaBuilder.buildType(param.type(), false)).setName(name)).build());
                         context.knownSecrets.add(name);
                     }
                 } else if (param.hasAnnotation(Config.class)) {
@@ -504,8 +478,8 @@ class FtlProcessor {
                     String name = param.annotation(Config.class).value().asString();
                     paramMappers.add(new VerbRegistry.ConfigSupplier(name, paramType));
                     if (!context.knownConfig.contains(name)) {
-                        context.moduleBuilder.addDecls(Decl.newBuilder().setConfig(xyz.block.ftl.v1.schema.Config.newBuilder()
-                                .setType(buildType(context, param.type(), false)).setName(name)));
+                        context.schemaBuilder.addDecls(Decl.newBuilder().setConfig(xyz.block.ftl.v1.schema.Config.newBuilder()
+                                .setType(context.schemaBuilder.buildType(param.type(), false)).setName(name)).build());
                         context.knownConfig.add(name);
                     }
                 } else if (context.knownTopics.containsKey(param.type().name())) {
@@ -550,13 +524,13 @@ class FtlProcessor {
             verbBuilder
                     .setName(verbName)
                     .setExport(exported)
-                    .setRequest(buildType(context, bodyParamType, exported))
-                    .setResponse(buildType(context, method.returnType(), exported));
+                    .setRequest(context.schemaBuilder.buildType(bodyParamType, exported))
+                    .setResponse(context.schemaBuilder.buildType(method.returnType(), exported));
 
             if (metadataCallback != null) {
                 metadataCallback.accept(verbBuilder);
             }
-            context.moduleBuilder
+            context.schemaBuilder
                     .addDecls(Decl.newBuilder().setVerb(verbBuilder)
                             .build());
 
@@ -656,184 +630,9 @@ class FtlProcessor {
                 launchMode.isAuxiliaryApplication(), !capabilities.isPresent(Capability.VERTX_WEBSOCKETS));
     }
 
-    private Type buildType(ExtractionContext context, org.jboss.jandex.Type type, boolean export) {
-        switch (type.kind()) {
-            case PRIMITIVE -> {
-                var prim = type.asPrimitiveType();
-                switch (prim.primitive()) {
-                    case INT, LONG, BYTE, SHORT -> {
-                        return Type.newBuilder().setInt(Int.newBuilder().build()).build();
-                    }
-                    case FLOAT, DOUBLE -> {
-                        return Type.newBuilder().setFloat(Float.newBuilder().build()).build();
-                    }
-                    case BOOLEAN -> {
-                        return Type.newBuilder().setBool(Bool.newBuilder().build()).build();
-                    }
-                    case CHAR -> {
-                        return Type.newBuilder().setString(xyz.block.ftl.v1.schema.String.newBuilder().build()).build();
-                    }
-                    default -> throw new RuntimeException("unknown primitive type: " + prim.primitive());
-                }
-            }
-            case VOID -> {
-                return Type.newBuilder().setUnit(Unit.newBuilder().build()).build();
-            }
-            case ARRAY -> {
-                ArrayType arrayType = type.asArrayType();
-                if (arrayType.componentType().kind() == org.jboss.jandex.Type.Kind.PRIMITIVE && arrayType
-                        .componentType().asPrimitiveType().primitive() == PrimitiveType.Primitive.BYTE) {
-                    return Type.newBuilder().setBytes(Bytes.newBuilder().build()).build();
-                }
-                return Type.newBuilder()
-                        .setArray(Array.newBuilder().setElement(buildType(context, arrayType.componentType(), export)).build())
-                        .build();
-            }
-            case CLASS -> {
-                var clazz = type.asClassType();
-                var info = context.index().getComputingIndex().getClassByName(clazz.name());
-
-                PrimitiveType unboxed = PrimitiveType.unbox(clazz);
-                if (unboxed != null) {
-                    Type primitive = buildType(context, unboxed, export);
-                    if (type.hasAnnotation(NOT_NULL)) {
-                        return primitive;
-                    }
-                    return Type.newBuilder().setOptional(Optional.newBuilder().setType(primitive)).build();
-                }
-                if (info != null && info.hasDeclaredAnnotation(GENERATED_REF)) {
-                    var ref = info.declaredAnnotation(GENERATED_REF);
-                    return Type.newBuilder()
-                            .setRef(Ref.newBuilder().setName(ref.value("name").asString())
-                                    .setModule(ref.value("module").asString()))
-                            .build();
-                }
-                if (clazz.name().equals(DotName.STRING_NAME)) {
-                    return Type.newBuilder().setString(xyz.block.ftl.v1.schema.String.newBuilder().build()).build();
-                }
-                if (clazz.name().equals(DotName.OBJECT_NAME) || clazz.name().equals(JSON_NODE)) {
-                    return Type.newBuilder().setAny(Any.newBuilder().build()).build();
-                }
-                if (clazz.name().equals(OFFSET_DATE_TIME)) {
-                    return Type.newBuilder().setTime(Time.newBuilder().build()).build();
-                }
-                if (clazz.name().equals(INSTANT)) {
-                    return Type.newBuilder().setTime(Time.newBuilder().build()).build();
-                }
-                if (clazz.name().equals(ZONED_DATE_TIME)) {
-                    return Type.newBuilder().setTime(Time.newBuilder().build()).build();
-                }
-                var existing = context.dataElements.get(new TypeKey(clazz.name().toString(), List.of()));
-                if (existing != null) {
-                    if (existing.exported() || !export || !existing.ref().getModule().equals(context.moduleName)) {
-                        return Type.newBuilder().setRef(existing.ref()).build();
-                    }
-                    //bit of an edge case, we have an existing non-exported object that we need to export
-                    for (var i = 0; i < context.moduleBuilder.getDeclsCount(); ++i) {
-                        var decl = context.moduleBuilder.getDecls(i);
-                        if (!decl.hasData()) {
-                            continue;
-                        }
-                        if (decl.getData().getName().equals(existing.ref().getName())) {
-                            context.moduleBuilder.setDecls(i,
-                                    decl.toBuilder().setData(decl.getData().toBuilder().setExport(true)).build());
-                            break;
-                        }
-                    }
-                    return Type.newBuilder().setRef(existing.ref()).build();
-                }
-                Data.Builder data = Data.newBuilder();
-                data.setName(clazz.name().local());
-                data.setExport(type.hasAnnotation(EXPORT) || export);
-                buildDataElement(context, data, clazz.name());
-                context.moduleBuilder.addDecls(Decl.newBuilder().setData(data).build());
-                Ref ref = Ref.newBuilder().setName(data.getName()).setModule(context.moduleName).build();
-                context.dataElements.put(new TypeKey(clazz.name().toString(), List.of()),
-                        new ExistingRef(ref, export || data.getExport()));
-                return Type.newBuilder().setRef(ref).build();
-            }
-            case PARAMETERIZED_TYPE -> {
-                var paramType = type.asParameterizedType();
-                if (paramType.name().equals(DotName.createSimple(List.class))) {
-                    return Type.newBuilder()
-                            .setArray(Array.newBuilder().setElement(buildType(context, paramType.arguments().get(0), export)))
-                            .build();
-                } else if (paramType.name().equals(DotName.createSimple(Map.class))) {
-                    return Type.newBuilder().setMap(xyz.block.ftl.v1.schema.Map.newBuilder()
-                            .setKey(buildType(context, paramType.arguments().get(0), export))
-                            .setValue(buildType(context, paramType.arguments().get(0), export)))
-                            .build();
-                } else if (paramType.name().equals(DotNames.OPTIONAL)) {
-                    //TODO: optional kinda sucks
-                    return Type.newBuilder()
-                            .setOptional(
-                                    Optional.newBuilder().setType(buildType(context, paramType.arguments().get(0), export)))
-                            .build();
-                } else if (paramType.name().equals(DotName.createSimple(HttpRequest.class))) {
-                    return Type.newBuilder()
-                            .setRef(Ref.newBuilder().setModule(BUILTIN).setName(HttpRequest.class.getSimpleName())
-                                    .addTypeParameters(buildType(context, paramType.arguments().get(0), export)))
-                            .build();
-                } else if (paramType.name().equals(DotName.createSimple(HttpResponse.class))) {
-                    return Type.newBuilder()
-                            .setRef(Ref.newBuilder().setModule(BUILTIN).setName(HttpResponse.class.getSimpleName())
-                                    .addTypeParameters(buildType(context, paramType.arguments().get(0), export))
-                                    .addTypeParameters(Type.newBuilder().setUnit(Unit.newBuilder().build())))
-                            .build();
-                } else {
-                    ClassInfo classByName = context.index().getComputingIndex().getClassByName(paramType.name());
-                    var cb = ClassType.builder(classByName.name());
-                    var main = buildType(context, cb.build(), export);
-                    var builder = main.toBuilder();
-                    var refBuilder = builder.getRef().toBuilder();
-
-                    for (var arg : paramType.arguments()) {
-                        refBuilder.addTypeParameters(buildType(context, arg, export));
-                    }
-                    builder.setRef(refBuilder);
-                    return builder.build();
-                }
-            }
-        }
-
-        throw new RuntimeException("NOT YET IMPLEMENTED");
-    }
-
-    private void buildDataElement(ExtractionContext context, Data.Builder data, DotName className) {
-        if (className == null || className.equals(DotName.OBJECT_NAME)) {
-            return;
-        }
-        var clazz = context.index.getComputingIndex().getClassByName(className);
-        if (clazz == null) {
-            return;
-        }
-        //TODO: handle getters and setters properly, also Jackson annotations etc
-        for (var field : clazz.fields()) {
-            if (!Modifier.isStatic(field.flags())) {
-                Field.Builder builder = Field.newBuilder().setName(field.name())
-                        .setType(buildType(context, field.type(), data.getExport()));
-                if (field.hasAnnotation(JsonAlias.class)) {
-                    var aliases = field.annotation(JsonAlias.class);
-                    if (aliases.value() != null) {
-                        for (var alias : aliases.value().asStringArray()) {
-                            builder.addMetadata(
-                                    Metadata.newBuilder().setAlias(MetadataAlias.newBuilder().setKind(0).setAlias(alias)));
-                        }
-                    }
-                }
-                data.addFields(builder.build());
-            }
-        }
-        buildDataElement(context, data, clazz.superName());
-    }
-
-    private record TypeKey(String name, List<String> typeParams) {
-
-    }
-
     record ExtractionContext(String moduleName, CombinedIndexBuildItem index, FTLRecorder recorder,
-            Module.Builder moduleBuilder,
-            Map<TypeKey, ExistingRef> dataElements, Set<String> knownSecrets, Set<String> knownConfig,
+            SchemaBuilder schemaBuilder,
+            Set<String> knownSecrets, Set<String> knownConfig,
             Map<DotName, TopicsBuildItem.DiscoveredTopic> knownTopics,
             Map<DotName, VerbClientBuildItem.DiscoveredClients> verbClients) {
     }
@@ -844,7 +643,4 @@ class FtlProcessor {
         REQUIRED
     }
 
-    record ExistingRef(Ref ref, boolean exported) {
-
-    }
 }
